requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pyyaml>=6.0
typer>=0.9.0
rich>=13.0
